AzureFunctions/requirements.txt
azure-functions
mysql-connector-python
bcrypt
python-dotenv
azure-storage-blob
azure-core